pyproject.toml
[project]
name = "MyBibliotheca"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.1",
    "gunicorn>=23.0.0",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "requests>=2.32.4",
    "wtforms>=3.1.2",
    "kuzu>=0.7.0",
]

[project.optional-dependencies]
migration = [
    "flask-sqlalchemy>=3.1.1",
]
